+++ Assets/_Scripts/old/MapGenerator.cs
@@ -1,17 +1,44 @@
+using System;
+using _Scripts._GradientNoise.OpenSimplex;
 using _Scripts.CellGeneration;
 using _Scripts.Helper;
 using UnityEngine;
 
 namespace _Scripts.old
 {
+    /**
+     * The type of Noise which should be used.
+     */
+    public enum NoiseType
+    {
+        PseudoRandom,
+        OpenSimplex,
+        Perlin,
+    }
+
+    /**
+     * This class controls the entire generation process.
+     * All required parameters are collected in it and passed to the corresponding methods.
+     */
     public class MapGenerator : MonoBehaviour
     {
-        // [SerializeField] private int width;
-        // [SerializeField] private int height;
+        [Header("General")] 
+        [SerializeField] public NoiseType noiseType;
         [SerializeField] private Vector2Int resolution = new Vector2Int(256, 144);
-
-        [SerializeField] private int smoothSteps = 5;
         [Range(0, 100)] [SerializeField] private int indoorsPercentage = 35;
+
+        // Seed
+        [Header("Seed")] public bool useRandomSeed = true;
+        [SerializeField] private string seed = "Hello World!";
+        [Range(1000.0f, 1000000.0f)] public float seedScale = 100000.0f;
+
+        // Gradient noise settings
+        [Header("Gradient Noise")] [Range(0.0f, 1.0f)]
+        public float noiseScale = 0.033f;
+
+        [Header("Pseudo Random")] [SerializeField]
+        private int smoothSteps = 7;
+
 
         // private properties
         private int[,] _valueMap;
@@ -20,82 +47,107 @@ namespace _Scripts.old
 
         void Start()
         {
-            // // init MapDisplay
+            // Initialization
             var mapPos = transform.position;
             _display = new MapDisplay(mapPos, resolution, gameObject);
 
-            // generate valueMap
-            _valueMap = GenerateMap();
-            SmoothMap();
+            // Generate and display valueMap
+            _valueMap = GenerateMap(seed, noiseType, seedScale, noiseScale, indoorsPercentage);
             _display.UpdateMapDisplay(_valueMap);
-
-            // init cell array
-            _cells = new Cell[resolution.x, resolution.y];
-
-            // init random generator
-            System.Random pRandom = new System.Random();
-
-            for (int x = 0; x < resolution.x; x++)
-            {
-                for (int y = 0; y < resolution.y; y++)
-                {
-                    // generate new cell
-                    Cell cell = new Cell(x, y);
-                    cell.CellIndex = new Vector2Int(x, y);
-
-                    // determine if cell is in- or outdoors
-                    cell.Indoors = _valueMap[x, y] == 1;
-
-                    // add cell to cells
-                    _cells[x, y] = cell;
-                }
-            }
         }
 
         void Update()
         {
-            // if (Input.GetMouseButtonDown(0))
-            // {
-            //     SmoothMap();
-            //
-            //     _display.UpdateMapDisplay(_valueMap);
-            // }
+            if (Input.GetMouseButtonDown(0) && noiseType == NoiseType.PseudoRandom)
+            {
+                SmoothMap();
+
+                _display.UpdateMapDisplay(_valueMap);
+            }
         }
 
-        private int[,] GenerateMap()
+        /**
+         * Generate the map.
+         */
+        private int[,] GenerateMap(string seedIn, NoiseType noiseTypeIn, float seedScaleIn, float noiseScaleIn,
+            int indoorsPercentageIn)
         {
-            // init result array
+            // Initialize result array
             int[,] result = new int[resolution.x, resolution.y];
-            System.Random prng = new System.Random();
 
-            // for (int x = 0; x < resolution.x; x++)
-            // {
-            //     for (int y = 0; y < resolution.y; y++)
-            //     {
-            //         if (x == 0 || x == resolution.x - 1 || y == 0 || y == resolution.y - 1)
-            //         {
-            //             result[x, y] = 1; // Wall
-            //         }
-            //         else
-            //         {
-            //             // Next(n) -> value between inclusive 0 and exclusive n
-            //             result[x, y] = prng.Next(101) < indoorsPercentage ? 1 : 0;
-            //         }
-            //     }
-            // }
+            float threshold;
+            double noiseValue;
 
-            for (int x = 0; x < resolution.x; x++)
+            // Get the coordinates
+            float seedOffset = seedIn.GetHashCode() / seedScaleIn;
+            float sampleX;
+            float sampleY;
+
+            switch (noiseTypeIn)
             {
-                for (int y = 0; y < resolution.y; y++)
-                {
-                    // Next(n) -> value between inclusive 0 and exclusive n
-                    result[x, y] = prng.Next(101) < indoorsPercentage ? 1 : 0;
-                }
+                case NoiseType.Perlin:
+
+                    for (int x = 0; x < resolution.x; x++)
+                    {
+                        for (int y = 0; y < resolution.y; y++)
+                        {
+                            // Get the coordinates
+                            sampleX = (x + seedOffset) * noiseScaleIn;
+                            sampleY = (y + seedOffset) * noiseScaleIn;
+
+                            threshold = Mathf.Lerp(0.0f, 1.0f, (float)indoorsPercentageIn / 100);
+                            noiseValue = Mathf.PerlinNoise(sampleX, sampleY);
+                            result[x, y] = noiseValue < threshold ? 1 : 0;
+                        }
+                    }
+
+                    break;
+
+                case NoiseType.OpenSimplex:
+
+                    for (int x = 0; x < resolution.x; x++)
+                    {
+                        for (int y = 0; y < resolution.y; y++)
+                        {
+                            // Get the coordinates
+                            sampleX = (x + seedOffset) * noiseScaleIn;
+                            sampleY = (y + seedOffset) * noiseScaleIn;
+
+                            OpenSimplexNoise openSimplexNoise = new OpenSimplexNoise(seedIn.GetHashCode());
+                            threshold = Mathf.Lerp(-1.0f, 1.0f, (float)indoorsPercentageIn / 100);
+                            noiseValue = (float)openSimplexNoise.Evaluate(sampleX, sampleY);
+                            result[x, y] = noiseValue < threshold ? 1 : 0;
+                        }
+                    }
+
+                    break;
+
+                case NoiseType.PseudoRandom:
+
+                    System.Random prng = new System.Random(seedIn.GetHashCode());
+
+                    for (int x = 0; x < resolution.x; x++)
+                    {
+                        for (int y = 0; y < resolution.y; y++)
+                        {
+                            // Next(n) -> value between inclusive 0 and exclusive n
+                            result[x, y] = prng.Next(101) < indoorsPercentageIn ? 1 : 0;
+                        }
+                    }
+
+                    break;
+
+                default:
+                    throw new ArgumentOutOfRangeException();
             }
+
 
             return result;
         }
 
+        /**
+         * Smooth the map.
+         */
         private void SmoothMap()
         {
             if (_valueMap == null) return;
@@ -116,26 +168,30 @@ namespace _Scripts.old
             }
         }
 
+        /**
+         * Apply the rules.
+         */
         private int ApplyRules(int xIndex, int yIndex)
         {
-            var neighbours = GetSurroundingWallCount(xIndex, yIndex);
+            var neighbours = GetLivingNeighboursCount(xIndex, yIndex);
 
-            if (neighbours > 4)
+            switch (neighbours)
             {
-                return 1; // Wall
+                case > 4:
+                    return 1;
+                case < 4:
+                    return 0;
+                default:
+                    return _valueMap[xIndex, yIndex];
             }
-
-            if (neighbours < 4)
-            {
-                return 0; // Empty tile
-            }
-
-            return _valueMap[xIndex, yIndex];
         }
 
-        private int GetSurroundingWallCount(int xIndex, int yIndex)
+        /**
+         * Get the number of neighbours with value 1.
+         */
+        private int GetLivingNeighboursCount(int xIndex, int yIndex)
         {
-            int wallCount = 0;
+            int neighbours = 0;
 
             for (int x = -1; x <= 1; x++)
             {
@@ -148,12 +204,12 @@ namespace _Scripts.old
                     if ((x == 0 && y == 0) || xPos < 0 || xPos >= resolution.x || yPos < 0 ||
                         yPos >= resolution.y) continue;
 
-                    // increment neighbours if wall
-                    if (_valueMap[xPos, yPos] == 1) wallCount++;
+                    // increment neighbours if 1
+                    if (_valueMap[xPos, yPos] == 1) neighbours++;
                 }
             }
 
-            return wallCount;
+            return neighbours;
         }
     }
 }