+++ Assets/_Scripts/old/ProceduralGeneration.cs
@@ -2,6 +2,9 @@ using UnityEngine;
 
 namespace _Scripts.old
 {
+    /**
+     * This class generates tiles on the x- and y-axis.
+     */
     public class ProceduralGeneration : MonoBehaviour
     {
         [SerializeField] private Vector2Int resolution = new Vector2Int(256, 144);
@@ -12,6 +15,7 @@ namespace _Scripts.old
             Generation();
         }
 
+        // The generation of the tiles
         private void Generation()
         {
             for (int x = 0; x < resolution.x; x++) // This will help to spawn a tile on the xAxis